Extract hook to a separate file
