+++ src/App.tsx
@@ -15,7 +15,7 @@ import { ScratchPad } from "./pages/ScratchPad"
 import Upgrade from "./pages/Upgrade"
 import Portfolio from "./pages/Portfolio"
 import { useSubscribeToContractEvent } from "./web3/hooks/useSubscribeToContractEvent"
-import { useContract } from "./web3/hooks/useERC20"
+import { useContract } from "./web3/hooks/useContract"
 import VendingMachineKeep from "@threshold-network/solidity-contracts/artifacts/VendingMachineKeep.json"
 import { useSubscribeToERC20TransferEvent } from "./web3/hooks/useSubscribeToERC20TransferEvent"
 import { Token } from "./enums"
